Added QgsProject, which uses QgsException.  QgsProject will eventually be responsible for reading and writing project files, which restore and save qgis state, respectively.
Note that now map layers are responsible for reading and restoring their
states to project files.  So have added readXML() and writeXML() members to
QgsMapLayer and QgsVectorLayer.  (Eventually QgsRasterLayer will have these,
too.)

Note that all of new project file semantics used only with command line
arguments.  If loaded from GUI, older methods used.  Sorta.  Given changes
rippled throughout code, the GUI "saveFile()" will also run some new code.

Incomplete implementation; currently only supports read of existing project
files and then only for vector layers.

Also eliminated "toInt()" members since casting should be done explicitly.
Rolled forward changes.

QgsPoint external inline members moved into class declaration, thus shortening
code.  Replaced pass by value with pass by reference to save unnecessary
copying.

In QgsRect, eliminated unnecessary copying.

Bugs: Updates main canvas window, but overview updated unnecessarily; and
overview map clipped incorrectly.


git-svn-id: http://svn.osgeo.org/qgis@2003 c8812cc2-4d05-0410-92ff-de0c093fc19c

+++ branches/qgsproject-branch/qgis/src/qgsmapcanvas.cpp
@@ -1012,7 +1012,10 @@ void QgsMapCanvas::mouseReleaseEvent(QMouseEvent * e)
           --it;
           --it;
           QgsPoint lastpoint = mCanvasProperties->coordXForm->transform(it->x(),it->y());
-          paint.drawLine(lastpoint.x(),lastpoint.y(),e->x(),e->y());
+          paint.drawLine(static_cast<int>(lastpoint.x()),
+                         static_cast<int>(lastpoint.y()),
+                         e->x(),
+                         e->y());
         }
         if(e->button()==Qt::RightButton)
         {
